requirements.txt
<<<<<<< HEAD
Flask==3.0.0
psycopg[binary]==3.2.2
python-dotenv==1.0.0
gunicorn==21.2.0
flasgger
=======
Flask
psycopg[binary]
python-dotenv
flasgger
flask-jwt-extended
werkzeug
>>>>>>> 2f24dcc (major changes)
